chore: update `.prettierignore`
