+++ packages/rolldown/rolldown-binding.wasi-browser.js
@@ -14,7 +14,7 @@ const __wasi = new __WASI({
   fs: __fs,
   preopens: {
     '/': '/',
-  },
+  }
 })
 
 const __emnapiContext = __emnapiGetDefaultContext()
@@ -36,13 +36,10 @@ const {
   asyncWorkPoolSize: 4,
   wasi: __wasi,
   onCreateWorker() {
-    const worker = new Worker(
-      new URL('./wasi-worker-browser.mjs', import.meta.url),
-      {
-        type: 'module',
-      },
-    )
-
+    const worker = new Worker(new URL('./wasi-worker-browser.mjs', import.meta.url), {
+      type: 'module',
+    })
+    
     worker.addEventListener('message', __wasmCreateOnMessageForFsProxy(__fs))
 
     return worker
@@ -62,17 +59,11 @@ const {
 })
 
 function __napi_rs_initialize_modules(__napiInstance) {
-  __napiInstance.exports[
-    '__napi_register__IsolatedDeclarationsResult_struct_0'
-  ]?.()
+  __napiInstance.exports['__napi_register__IsolatedDeclarationsResult_struct_0']?.()
   __napiInstance.exports['__napi_register__isolated_declaration_1']?.()
-  __napiInstance.exports[
-    '__napi_register__TypeScriptBindingOptions_struct_2'
-  ]?.()
+  __napiInstance.exports['__napi_register__TypeScriptBindingOptions_struct_2']?.()
   __napiInstance.exports['__napi_register__ReactBindingOptions_struct_3']?.()
-  __napiInstance.exports[
-    '__napi_register__ArrowFunctionsBindingOptions_struct_4'
-  ]?.()
+  __napiInstance.exports['__napi_register__ArrowFunctionsBindingOptions_struct_4']?.()
   __napiInstance.exports['__napi_register__ES2015BindingOptions_struct_5']?.()
   __napiInstance.exports['__napi_register__TransformOptions_struct_6']?.()
   __napiInstance.exports['__napi_register__Sourcemap_struct_7']?.()
@@ -87,46 +78,24 @@ function __napi_rs_initialize_modules(__napiInstance) {
   __napiInstance.exports['__napi_register__BindingOutputOptions_struct_10']?.()
   __napiInstance.exports['__napi_register__BindingPluginContext_struct_11']?.()
   __napiInstance.exports['__napi_register__BindingPluginContext_impl_17']?.()
-  __napiInstance.exports[
-    '__napi_register__BindingPluginContextResolvedId_struct_18'
-  ]?.()
+  __napiInstance.exports['__napi_register__BindingPluginContextResolvedId_struct_18']?.()
   __napiInstance.exports['__napi_register__BindingPluginOptions_struct_19']?.()
-  __napiInstance.exports[
-    '__napi_register__BindingPluginWithIndex_struct_20'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingTransformPluginContext_struct_21'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingTransformPluginContext_impl_23'
-  ]?.()
+  __napiInstance.exports['__napi_register__BindingPluginWithIndex_struct_20']?.()
+  __napiInstance.exports['__napi_register__BindingTransformPluginContext_struct_21']?.()
+  __napiInstance.exports['__napi_register__BindingTransformPluginContext_impl_23']?.()
   __napiInstance.exports['__napi_register__BindingAssetSource_struct_24']?.()
   __napiInstance.exports['__napi_register__BindingEmittedAsset_struct_25']?.()
   __napiInstance.exports['__napi_register__BindingHookLoadOutput_struct_26']?.()
-  __napiInstance.exports[
-    '__napi_register__BindingHookRenderChunkOutput_struct_27'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingHookResolveIdExtraOptions_struct_28'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingHookResolveIdOutput_struct_29'
-  ]?.()
+  __napiInstance.exports['__napi_register__BindingHookRenderChunkOutput_struct_27']?.()
+  __napiInstance.exports['__napi_register__BindingHookResolveIdExtraOptions_struct_28']?.()
+  __napiInstance.exports['__napi_register__BindingHookResolveIdOutput_struct_29']?.()
   __napiInstance.exports['__napi_register__BindingHookSideEffects_30']?.()
-  __napiInstance.exports[
-    '__napi_register__BindingHookTransformOutput_struct_31'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingPluginContextResolveOptions_struct_32'
-  ]?.()
+  __napiInstance.exports['__napi_register__BindingHookTransformOutput_struct_31']?.()
+  __napiInstance.exports['__napi_register__BindingPluginContextResolveOptions_struct_32']?.()
   __napiInstance.exports['__napi_register__BindingBuiltinPlugin_struct_33']?.()
   __napiInstance.exports['__napi_register__BindingBuiltinPluginName_34']?.()
-  __napiInstance.exports[
-    '__napi_register__ParallelJsPluginRegistry_struct_35'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__ParallelJsPluginRegistry_impl_37'
-  ]?.()
+  __napiInstance.exports['__napi_register__ParallelJsPluginRegistry_struct_35']?.()
+  __napiInstance.exports['__napi_register__ParallelJsPluginRegistry_impl_37']?.()
   __napiInstance.exports['__napi_register__register_plugins_38']?.()
   __napiInstance.exports['__napi_register__BindingLog_struct_39']?.()
   __napiInstance.exports['__napi_register__BindingLogLevel_40']?.()
@@ -152,16 +121,12 @@ export const BindingOutputAsset = __napiModule.exports.BindingOutputAsset
 export const BindingOutputChunk = __napiModule.exports.BindingOutputChunk
 export const BindingOutputs = __napiModule.exports.BindingOutputs
 export const BindingPluginContext = __napiModule.exports.BindingPluginContext
-export const BindingTransformPluginContext =
-  __napiModule.exports.BindingTransformPluginContext
+export const BindingTransformPluginContext = __napiModule.exports.BindingTransformPluginContext
 export const Bundler = __napiModule.exports.Bundler
 export const FinalBindingOutputs = __napiModule.exports.FinalBindingOutputs
-export const ParallelJsPluginRegistry =
-  __napiModule.exports.ParallelJsPluginRegistry
-export const BindingBuiltinPluginName =
-  __napiModule.exports.BindingBuiltinPluginName
-export const BindingHookSideEffects =
-  __napiModule.exports.BindingHookSideEffects
+export const ParallelJsPluginRegistry = __napiModule.exports.ParallelJsPluginRegistry
+export const BindingBuiltinPluginName = __napiModule.exports.BindingBuiltinPluginName
+export const BindingHookSideEffects = __napiModule.exports.BindingHookSideEffects
 export const BindingLogLevel = __napiModule.exports.BindingLogLevel
 export const isolatedDeclaration = __napiModule.exports.isolatedDeclaration
 export const registerPlugins = __napiModule.exports.registerPlugins
