+++ packages/rolldown/rolldown-binding.wasi.cjs
@@ -21,7 +21,7 @@ const __wasi = new __nodeWASI({
   env: process.env,
   preopens: {
     [__rootDir]: __rootDir,
-  },
+  }
 })
 
 const __emnapiContext = __emnapiGetDefaultContext()
@@ -32,14 +32,8 @@ const __sharedMemory = new WebAssembly.Memory({
   shared: true,
 })
 
-let __wasmFilePath = __nodePath.join(
-  __dirname,
-  'rolldown-binding.wasm32-wasi.wasm',
-)
-const __wasmDebugFilePath = __nodePath.join(
-  __dirname,
-  'rolldown-binding.wasm32-wasi.debug.wasm',
-)
+let __wasmFilePath = __nodePath.join(__dirname, 'rolldown-binding.wasm32-wasi.wasm')
+const __wasmDebugFilePath = __nodePath.join(__dirname, 'rolldown-binding.wasm32-wasi.debug.wasm')
 
 if (__nodeFs.existsSync(__wasmDebugFilePath)) {
   __wasmFilePath = __wasmDebugFilePath
@@ -47,23 +41,14 @@ if (__nodeFs.existsSync(__wasmDebugFilePath)) {
   try {
     __wasmFilePath = __nodePath.resolve('@rolldown/binding-wasm32-wasi')
   } catch {
-    throw new Error(
-      'Cannot find rolldown-binding.wasm32-wasi.wasm file, and @rolldown/binding-wasm32-wasi package is not installed.',
-    )
+    throw new Error('Cannot find rolldown-binding.wasm32-wasi.wasm file, and @rolldown/binding-wasm32-wasi package is not installed.')
   }
 }
 
-const {
-  instance: __napiInstance,
-  module: __wasiModule,
-  napiModule: __napiModule,
-} = __emnapiInstantiateNapiModuleSync(__nodeFs.readFileSync(__wasmFilePath), {
+const { instance: __napiInstance, module: __wasiModule, napiModule: __napiModule } = __emnapiInstantiateNapiModuleSync(__nodeFs.readFileSync(__wasmFilePath), {
   context: __emnapiContext,
-  asyncWorkPoolSize: (function () {
-    const threadsSizeFromEnv = Number(
-      process.env.NAPI_RS_ASYNC_WORK_POOL_SIZE ??
-        process.env.UV_THREADPOOL_SIZE,
-    )
+  asyncWorkPoolSize: (function() {
+    const threadsSizeFromEnv = Number(process.env.NAPI_RS_ASYNC_WORK_POOL_SIZE ?? process.env.UV_THREADPOOL_SIZE)
     // NaN > 0 is false
     if (threadsSizeFromEnv > 0) {
       return threadsSizeFromEnv
@@ -93,21 +78,15 @@ const {
   },
   beforeInit({ instance }) {
     __napi_rs_initialize_modules(instance)
-  },
+  }
 })
 
 function __napi_rs_initialize_modules(__napiInstance) {
-  __napiInstance.exports[
-    '__napi_register__IsolatedDeclarationsResult_struct_0'
-  ]?.()
+  __napiInstance.exports['__napi_register__IsolatedDeclarationsResult_struct_0']?.()
   __napiInstance.exports['__napi_register__isolated_declaration_1']?.()
-  __napiInstance.exports[
-    '__napi_register__TypeScriptBindingOptions_struct_2'
-  ]?.()
+  __napiInstance.exports['__napi_register__TypeScriptBindingOptions_struct_2']?.()
   __napiInstance.exports['__napi_register__ReactBindingOptions_struct_3']?.()
-  __napiInstance.exports[
-    '__napi_register__ArrowFunctionsBindingOptions_struct_4'
-  ]?.()
+  __napiInstance.exports['__napi_register__ArrowFunctionsBindingOptions_struct_4']?.()
   __napiInstance.exports['__napi_register__ES2015BindingOptions_struct_5']?.()
   __napiInstance.exports['__napi_register__TransformOptions_struct_6']?.()
   __napiInstance.exports['__napi_register__Sourcemap_struct_7']?.()
@@ -122,46 +101,24 @@ function __napi_rs_initialize_modules(__napiInstance) {
   __napiInstance.exports['__napi_register__BindingOutputOptions_struct_10']?.()
   __napiInstance.exports['__napi_register__BindingPluginContext_struct_11']?.()
   __napiInstance.exports['__napi_register__BindingPluginContext_impl_17']?.()
-  __napiInstance.exports[
-    '__napi_register__BindingPluginContextResolvedId_struct_18'
-  ]?.()
+  __napiInstance.exports['__napi_register__BindingPluginContextResolvedId_struct_18']?.()
   __napiInstance.exports['__napi_register__BindingPluginOptions_struct_19']?.()
-  __napiInstance.exports[
-    '__napi_register__BindingPluginWithIndex_struct_20'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingTransformPluginContext_struct_21'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingTransformPluginContext_impl_23'
-  ]?.()
+  __napiInstance.exports['__napi_register__BindingPluginWithIndex_struct_20']?.()
+  __napiInstance.exports['__napi_register__BindingTransformPluginContext_struct_21']?.()
+  __napiInstance.exports['__napi_register__BindingTransformPluginContext_impl_23']?.()
   __napiInstance.exports['__napi_register__BindingAssetSource_struct_24']?.()
   __napiInstance.exports['__napi_register__BindingEmittedAsset_struct_25']?.()
   __napiInstance.exports['__napi_register__BindingHookLoadOutput_struct_26']?.()
-  __napiInstance.exports[
-    '__napi_register__BindingHookRenderChunkOutput_struct_27'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingHookResolveIdExtraOptions_struct_28'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingHookResolveIdOutput_struct_29'
-  ]?.()
+  __napiInstance.exports['__napi_register__BindingHookRenderChunkOutput_struct_27']?.()
+  __napiInstance.exports['__napi_register__BindingHookResolveIdExtraOptions_struct_28']?.()
+  __napiInstance.exports['__napi_register__BindingHookResolveIdOutput_struct_29']?.()
   __napiInstance.exports['__napi_register__BindingHookSideEffects_30']?.()
-  __napiInstance.exports[
-    '__napi_register__BindingHookTransformOutput_struct_31'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__BindingPluginContextResolveOptions_struct_32'
-  ]?.()
+  __napiInstance.exports['__napi_register__BindingHookTransformOutput_struct_31']?.()
+  __napiInstance.exports['__napi_register__BindingPluginContextResolveOptions_struct_32']?.()
   __napiInstance.exports['__napi_register__BindingBuiltinPlugin_struct_33']?.()
   __napiInstance.exports['__napi_register__BindingBuiltinPluginName_34']?.()
-  __napiInstance.exports[
-    '__napi_register__ParallelJsPluginRegistry_struct_35'
-  ]?.()
-  __napiInstance.exports[
-    '__napi_register__ParallelJsPluginRegistry_impl_37'
-  ]?.()
+  __napiInstance.exports['__napi_register__ParallelJsPluginRegistry_struct_35']?.()
+  __napiInstance.exports['__napi_register__ParallelJsPluginRegistry_impl_37']?.()
   __napiInstance.exports['__napi_register__register_plugins_38']?.()
   __napiInstance.exports['__napi_register__BindingLog_struct_39']?.()
   __napiInstance.exports['__napi_register__BindingLogLevel_40']?.()
@@ -187,16 +144,12 @@ module.exports.BindingOutputAsset = __napiModule.exports.BindingOutputAsset
 module.exports.BindingOutputChunk = __napiModule.exports.BindingOutputChunk
 module.exports.BindingOutputs = __napiModule.exports.BindingOutputs
 module.exports.BindingPluginContext = __napiModule.exports.BindingPluginContext
-module.exports.BindingTransformPluginContext =
-  __napiModule.exports.BindingTransformPluginContext
+module.exports.BindingTransformPluginContext = __napiModule.exports.BindingTransformPluginContext
 module.exports.Bundler = __napiModule.exports.Bundler
 module.exports.FinalBindingOutputs = __napiModule.exports.FinalBindingOutputs
-module.exports.ParallelJsPluginRegistry =
-  __napiModule.exports.ParallelJsPluginRegistry
-module.exports.BindingBuiltinPluginName =
-  __napiModule.exports.BindingBuiltinPluginName
-module.exports.BindingHookSideEffects =
-  __napiModule.exports.BindingHookSideEffects
+module.exports.ParallelJsPluginRegistry = __napiModule.exports.ParallelJsPluginRegistry
+module.exports.BindingBuiltinPluginName = __napiModule.exports.BindingBuiltinPluginName
+module.exports.BindingHookSideEffects = __napiModule.exports.BindingHookSideEffects
 module.exports.BindingLogLevel = __napiModule.exports.BindingLogLevel
 module.exports.isolatedDeclaration = __napiModule.exports.isolatedDeclaration
 module.exports.registerPlugins = __napiModule.exports.registerPlugins
