+++ packages/rolldown/src/wasi-worker-browser.mjs
@@ -1,4 +1,9 @@
-import { instantiateNapiModuleSync, MessageHandler, WASI, createFsProxy } from '@napi-rs/wasm-runtime'
+import {
+  instantiateNapiModuleSync,
+  MessageHandler,
+  WASI,
+  createFsProxy,
+} from '@napi-rs/wasm-runtime'
 import { memfsExported as __memfsExported } from '@napi-rs/wasm-runtime/fs'
 
 const fs = createFsProxy(__memfsExported)
@@ -14,7 +19,7 @@ const handler = new MessageHandler({
         // eslint-disable-next-line no-console
         console.log.apply(console, arguments)
       },
-      printErr: function() {
+      printErr: function () {
         // eslint-disable-next-line no-console
         console.error.apply(console, arguments)
       },
