+++ packages/rolldown/wasi-worker.mjs
@@ -1,21 +1,17 @@
-import fs from 'node:fs'
-import { createRequire } from 'node:module'
-import { parse } from 'node:path'
-import { WASI } from 'node:wasi'
-import { parentPort, Worker } from 'node:worker_threads'
+import fs from "node:fs";
+import { createRequire } from "node:module";
+import { parse } from "node:path";
+import { WASI } from "node:wasi";
+import { parentPort, Worker } from "node:worker_threads";
 
-const require = createRequire(import.meta.url)
+const require = createRequire(import.meta.url);
 
-const {
-  instantiateNapiModuleSync,
-  MessageHandler,
-  getDefaultContext,
-} = require('@napi-rs/wasm-runtime')
+const { instantiateNapiModuleSync, MessageHandler, getDefaultContext } = require("@napi-rs/wasm-runtime");
 
 if (parentPort) {
-  parentPort.on('message', (data) => {
-    globalThis.onmessage({ data })
-  })
+  parentPort.on("message", (data) => {
+    globalThis.onmessage({ data });
+  });
 }
 
 Object.assign(globalThis, {
@@ -23,18 +19,18 @@ Object.assign(globalThis, {
   require,
   Worker,
   importScripts: function (f) {
-    ;(0, eval)(fs.readFileSync(f, 'utf8') + '//# sourceURL=' + f)
+    ;(0, eval)(fs.readFileSync(f, "utf8") + "//# sourceURL=" + f);
   },
   postMessage: function (msg) {
     if (parentPort) {
-      parentPort.postMessage(msg)
+      parentPort.postMessage(msg);
     }
   },
-})
+});
 
-const emnapiContext = getDefaultContext()
+const emnapiContext = getDefaultContext();
 
-const __rootDir = parse(process.cwd()).root
+const __rootDir = parse(process.cwd()).root;
 
 const handler = new MessageHandler({
   onLoad({ wasmModule, wasmMemory }) {
@@ -44,7 +40,7 @@ const handler = new MessageHandler({
       preopens: {
         [__rootDir]: __rootDir,
       },
-    })
+    });
 
     return instantiateNapiModuleSync(wasmModule, {
       childThread: true,
@@ -55,13 +51,13 @@ const handler = new MessageHandler({
           ...importObject.env,
           ...importObject.napi,
           ...importObject.emnapi,
-          memory: wasmMemory,
-        }
+          memory: wasmMemory
+        };
       },
-    })
+    });
   },
-})
+});
 
 globalThis.onmessage = function (e) {
-  handler.handle(e)
-}
+  handler.handle(e);
+};
